lv-preisrechner/backend/pyproject.toml
[project]
name = "lv-preisrechner"
version = "0.1.0"
description = "LV-Preisrechner für Trockenbauer — Upload, DNA-Matching, ausgefülltes PDF zurück"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "psycopg2-binary>=2.9.10",
  "pydantic[email]>=2.9.0",
  "pydantic-settings>=2.6.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<4.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.17",
  "anthropic>=0.40.0",
  "pdfplumber>=0.11.4",
  "pymupdf>=1.24.14",
  "pillow>=11.0.0",
  "structlog>=24.4.0",
  "httpx>=0.28.0",
  "alembic>=1.14.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "httpx>=0.28.0",
  "ruff>=0.7.0",
  "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
